requirements-doc.txt
jupyterlab==4.0.9
jupyter_sphinx==0.4.0
numpydoc==1.6.0
pydata-sphinx-theme==0.14.4
pyvista==0.43.1
sphinx==7.2.6
sphinx-design==0.5.0
sphinx-gallery==0.15.0
trame==3.3.0
trame-vtk==2.6.2
trame-vuetify==2.3.1
